+++ plugins/postcss-lab-function/src/convert-lch-to-display-p3.ts
@@ -1,5 +1,4 @@
-import { D50_to_D65, gam_P3, Lab_to_XYZ, LCH_to_Lab, lin_P3, lin_P3_to_XYZ, OKLab_to_OKLCH, OKLab_to_XYZ, OKLCH_to_OKLab, XYZ_to_lin_P3, XYZ_to_OKLab } from './css-color-4/conversions.js';
-import { clip, inGamut, mapGamut } from './css-color-4/map-gamut';
+import { calculations, conversions, utils } from '@csstools/color-helpers';
 
 type color = [number, number, number];
 
@@ -14,16 +13,16 @@ export function lchToDisplayP3(lchRaw: color): [color, boolean] {
 	const lch = [lchL, lchCRaw, lchHRaw % 360] as color;
 
 	let conversion = lch;
-	conversion = LCH_to_Lab(conversion);
+	conversion = conversions.LCH_to_Lab(conversion);
 
 	// https://www.w3.org/TR/css-color-4/#oklab-lab-to-predefined
 	// 1. Convert Lab to(D50 - adapted) XYZ
-	conversion = Lab_to_XYZ(conversion);
+	conversion = conversions.Lab_to_XYZ(conversion);
 
 	let oklch = conversion.slice() as color;
-	oklch = D50_to_D65(oklch);
-	oklch = XYZ_to_OKLab(oklch);
-	oklch = OKLab_to_OKLCH(oklch);
+	oklch = conversions.D50_to_D65(oklch);
+	oklch = conversions.XYZ_to_OKLab(oklch);
+	oklch = conversions.OKLab_to_OKLCH(oklch);
 	if (oklch[0] < 0.000001) {
 		oklch = [0, 0, 0] as color;
 	}
@@ -33,25 +32,25 @@ export function lchToDisplayP3(lchRaw: color): [color, boolean] {
 	}
 
 	// 2. If needed, convert from a D50 whitepoint(used by Lab) to the D65 whitepoint used in sRGB and most other RGB spaces, with the Bradford transform.prophoto - rgb' does not require this step.
-	conversion = D50_to_D65(conversion);
+	conversion = conversions.D50_to_D65(conversion);
 	// 3. Convert from(D65 - adapted) CIE XYZ to linear P3
-	conversion = XYZ_to_lin_P3(conversion);
+	conversion = conversions.XYZ_to_lin_P3(conversion);
 	// 4. Convert from linear - light P3 to P3(do gamma encoding)
-	conversion = gam_P3(conversion);
+	conversion = conversions.gam_P3(conversion);
 
-	if (inGamut(conversion)) {
-		return [clip(conversion), true];
+	if (utils.inGamut(conversion)) {
+		return [utils.clip(conversion), true];
 	}
 
-	return [mapGamut(oklch, (x: color) => {
-		x = OKLCH_to_OKLab(x);
-		x = OKLab_to_XYZ(x);
-		x = XYZ_to_lin_P3(x);
-		return gam_P3(x);
+	return [calculations.mapGamut(oklch, (x: color) => {
+		x = conversions.OKLCH_to_OKLab(x);
+		x = conversions.OKLab_to_XYZ(x);
+		x = conversions.XYZ_to_lin_P3(x);
+		return conversions.gam_P3(x);
 	}, (x: color) => {
-		x = lin_P3(x);
-		x = lin_P3_to_XYZ(x);
-		x = XYZ_to_OKLab(x);
-		return OKLab_to_OKLCH(x);
+		x = conversions.lin_P3(x);
+		x = conversions.lin_P3_to_XYZ(x);
+		x = conversions.XYZ_to_OKLab(x);
+		return conversions.OKLab_to_OKLCH(x);
 	}), false];
 }