+++ rollup/configs/externals.mjs
@@ -10,6 +10,7 @@ export const externalsForCLI = [
 	'postcss-html',
 
 	'@csstools/cascade-layer-name-parser',
+	'@csstools/color-helpers',
 	'@csstools/css-parser-algorithms',
 	'@csstools/css-tokenizer',
 	'@csstools/media-query-list-parser',
@@ -87,6 +88,7 @@ export const externalsForPlugin = [
 	'postcss-html',
 
 	'@csstools/cascade-layer-name-parser',
+	'@csstools/color-helpers',
 	'@csstools/css-parser-algorithms',
 	'@csstools/css-tokenizer',
 	'@csstools/media-query-list-parser',